Remove unnecessary use of `read-string` in Middle Character
For some reason `read-string` was being called to parse a string to a string in Middle Character, along with a totally unnecessary try/catch block.

This cleans that up.

It looks like most if not all of the PSB2 problems have this issue and we'll need to go through and update all of those to not use `read-string`.