Further improve Middle Character and Twitter
Upon further inspection, Richard and I realized that we didn't need to have `parsed-outputs` at all in problems that just return a string (take the top value from the `:string` stack).

So we removed all that from the Middle Character and Twitter problems, and removed the unnecessary calls to `str` in the `levenshtein-distance` calls.

Co-authored-by: RichardLussier <lussi036@morris.umn.edu>