Refactor keyword approach to downsampling with functions
Specifying the type of downsampling via keywords (e.g. `:ds-function` and `:case-rand`) was very fragile and requires making changes to the core of the system whenever you wanted to change the approach to downsampling.

This commit changes that to instead have a `downsample` function specified in `argmap`. If no `downsample` function is specified, it defaults to using the `training-data` unchanged, i.e., no downsampling.

This arguably isn't "done done" because we still have to have the `downsample?` entry. If you set the `downsample` entry in `argmap`, but forget to set `downsample?` to `true`, things break, which is really annoying and Not Good. Unfortunately `downsample?` is used in several other places in `gp-loop` in ways I don't really understand at the moment, and until those can be changed we're stuck keeping `downsample?` for now.